% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_v2.R
\name{assemble}
\alias{assemble}
\title{Align Homologous Exons}
\usage{
assemble(map, data)
}
\arguments{
\item{key}{A file containing homologous exons between species. 
data A list containing INCLUSION files from vast-tools ouput. The order of species in list should match the exon 'key'}
}
\description{
Parse and assemble AS data between species
}
\examples{
assemble(map = AltEx.key, data = list(INCLUSION_LEVELS_FULL.Hsa, INCLUSION_LEVELS_FULL.Ptr, INCLUSION_LEVELS_FULL.Mma)) -> PSI_QS_FULL
}
