% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_v2.R
\name{filter_exons}
\alias{filter_exons}
\title{Filter Exons with Zero Variance}
\usage{
filter_exons(data, sp, indivs, stringent = FALSE)
}
\arguments{
\item{data}{The PSI file to be filtered
sp The total number of columns containing exon names
indivs A numerical vector containing the number of individuals for each tissue, in order from left to right
stringent Applies a stronger filter to return exons w/ at least 2 individuals per tissue/species (Required for EVE model). Defaults to FALSE.}
}
\description{
Removes exons without evidence of alternative splicing
}
\examples{
filter_exons(data = PSI_brain, sp = 3, indivs = c(6,6,3), stringent = TRUE) -> PSI_brain

# don't forget to subset the complementary varPSI file to match
varPSI_brain = varPSI_brain[varPSI_brain$V1 \%in\% PSI_brain$V1, ]
}
