% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize_v2.R
\name{finalize}
\alias{finalize}
\title{Finalize PSI and varPSI files}
\usage{
finalize(data, sp, PSI = TRUE)
}
\arguments{
\item{data}{File created by filter_exons() to be formatted for EVE model.
sp Total number of columns w/ exon names.
PSI Is the file you are finalizing a PSI file? Defaults to TRUE.}
}
\description{
Finalizes PSI and varPSI files, to be handled by EVE. It changes PSI values to 0-1 scale and all NA's are replaced by -1. Also, any extra columns w/ exon names are removed.
}
\examples{
finalize(data = PSI_brain, sp = 3, PSI = TRUE) -> PSI_brain
finalize(data = varPSI_brain, sp = 3, PSI = FALSE) -> varPSI_brain
}
