% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_v2.R
\name{separate}
\alias{separate}
\title{Separate PSI and QS data}
\usage{
separate(data, sp, tiss)
}
\arguments{
\item{data}{The file created by 'align' function, containing all species' PSI and Quality Score information
sp The total number of columns containing exon names
tiss A character vector containing tissue names (not case-sensitive, but spelling counts)}
}
\description{
Retrieves PSI data and calculates var(PSI) between all tissues.
}
\examples{
separate(data = PSI_QS_FULL, sp = 3, tiss = c("brain", "cerebellum", "heart", "kidney", "liver", "testis"))
}
